Create README.md
This morning, the organization experienced a Distributed Denial of Service (DDoS) attack, where an overwhelming flood of ICMP packets targeted the network. This attack rendered the internal network inaccessible for two hours, impacting all critical network services and internal traffic. The incident was caused by a vulnerability in the network firewall, which lacked proper configuration to filter ICMP packets. The attack disrupted business operations, resulting in downtime and a potential loss of productivity.